rename use-register
